half-working auth prompts
