+++ qutebrowser/mainwindow/statusbar/bar.py
@@ -25,8 +25,7 @@ from PyQt5.QtWidgets import QWidget, QHBoxLayout, QStackedLayout, QSizePolicy
 from qutebrowser.config import config, style
 from qutebrowser.utils import usertypes, log, objreg, utils
 from qutebrowser.mainwindow.statusbar import (command, progress, keystring,
-                                              percentage, url, prompt,
-                                              tabindex)
+                                              percentage, url, tabindex)
 from qutebrowser.mainwindow.statusbar import text as textwidget
